refactor nb_device bits into its own module
