+++ device/src/nb_device/test/mod.rs
@@ -3,7 +3,7 @@ mod util;
 use crate::test_util::*;
 use util::*;
 
-use crate::Event;
+use crate::nb_device::Event;
 #[test]
 fn test_join_rx1() {
     let mut device = test_device();
@@ -14,7 +14,7 @@ fn test_join_rx1() {
     assert!(matches!(response, Response::TimeoutRequest(5100)));
     device.get_radio().set_rxtx_handler(handle_join_request);
     // send a radio event to let the radio device indicate a packet was received
-    let response = device.handle_event(Event::RadioEvent(radio::Event::PhyEvent(()))).unwrap();
+    let response = device.handle_event(Event::RadioEvent(radio::Event::Phy(()))).unwrap();
     assert!(matches!(response, Response::JoinSuccess));
     assert!(device.get_session_keys().is_some());
 }
@@ -34,7 +34,7 @@ fn test_join_rx2() {
     let response = device.handle_event(Event::TimeoutFired).unwrap();
     assert!(matches!(response, Response::TimeoutRequest(6100)));
     // send a radio event to let the radio device indicate a packet was received
-    let response = device.handle_event(Event::RadioEvent(radio::Event::PhyEvent(()))).unwrap();
+    let response = device.handle_event(Event::RadioEvent(radio::Event::Phy(()))).unwrap();
     assert!(matches!(response, Response::JoinSuccess));
     assert!(device.get_session_keys().is_some());
 }
@@ -82,7 +82,7 @@ fn test_confirmed_uplink_with_ack_rx1() {
     assert!(matches!(response, Response::TimeoutRequest(1100)));
     device.get_radio().set_rxtx_handler(handle_data_uplink_with_link_adr_req);
     // send a radio event to let the radio device indicate a packet was received
-    let response = device.handle_event(Event::RadioEvent(radio::Event::PhyEvent(()))).unwrap();
+    let response = device.handle_event(Event::RadioEvent(radio::Event::Phy(()))).unwrap();
     assert!(matches!(response, Response::DownlinkReceived(0)));
 }
 
@@ -101,7 +101,7 @@ fn test_confirmed_uplink_with_ack_rx2() {
     assert!(matches!(response, Response::TimeoutRequest(2100)));
     device.get_radio().set_rxtx_handler(handle_data_uplink_with_link_adr_req);
     // send a radio event to let the radio device indicate a packet was received
-    let response = device.handle_event(Event::RadioEvent(radio::Event::PhyEvent(()))).unwrap();
+    let response = device.handle_event(Event::RadioEvent(radio::Event::Phy(()))).unwrap();
     assert!(matches!(response, Response::DownlinkReceived(0)));
 }
 
@@ -116,7 +116,7 @@ fn test_link_adr_ans() {
     assert!(matches!(response, Response::TimeoutRequest(1100)));
     device.get_radio().set_rxtx_handler(handle_data_uplink_with_link_adr_req);
     // send a radio event to let the radio device indicate a packet was received
-    let response = device.handle_event(Event::RadioEvent(radio::Event::PhyEvent(()))).unwrap();
+    let response = device.handle_event(Event::RadioEvent(radio::Event::Phy(()))).unwrap();
     assert!(matches!(response, Response::DownlinkReceived(0)));
     // send another uplink which should carry the LinkAdrAns
     let response = device.send(&[0; 1], 1, true).unwrap();
@@ -125,6 +125,6 @@ fn test_link_adr_ans() {
     assert!(matches!(response, Response::TimeoutRequest(1100)));
     device.get_radio().set_rxtx_handler(handle_data_uplink_with_link_adr_ans);
     // send a radio event to let the radio device indicate a packet was received
-    let response = device.handle_event(Event::RadioEvent(radio::Event::PhyEvent(()))).unwrap();
+    let response = device.handle_event(Event::RadioEvent(radio::Event::Phy(()))).unwrap();
     assert!(matches!(response, Response::DownlinkReceived(1)));
 }
